fix refreshToken
reworked refresh token logic
fixed errors
