+++ src/config/backend.js
@@ -1,0 +1,2 @@
+export const BACKEND_HOST = "http://localhost:3000";
+//export const BACKEND_HOST = "https://projectpilotbackend.onrender.com";
